Hashtag and KeyWord services - Complete

+++ src/main/java/us/kanddys/laia/modules/ecommerce/controller/HashtagController.java
@@ -18,7 +18,7 @@ public class HashtagController {
    private HashtagProductService hashtagProductService;
 
    @MutationMapping("cHashtag")
-   public Integer createHashtag(@Argument String value) {
+   public Long createHashtag(@Argument String value) {
       return hashtagService.createHashtag(value);
    }
